hooks moved from utils to hooks folder
