+++ assets/js/features/goals/GoalForm/useForm.tsx
@@ -8,7 +8,7 @@ import * as Timeframes from "@/utils/timeframes";
 
 import { useNavigateTo } from "@/routes/useNavigateTo";
 import { useNavigate } from "react-router-dom";
-import { useListState } from "@/utils/useListState";
+import { useListState } from "@/hooks/useListState";
 import { Paths } from "@/routes/paths";
 import { PermissionsState } from "@/features/Permissions/usePermissionsState";
 import { usePermissionsState } from "@/features/Permissions/usePermissionsState";
